+++ instapy2/like_util.py
@@ -12,12 +12,12 @@ class LikeUtil:
 
     def get_medias_for_tag(self, tag: str = None, amount: int = 50, skip_top_posts: bool = False, randomize: bool = False, media_type: MediaType = None) -> list[Media]:
         # media_type is unused currently
-        if media_type is None:
-            media_type = [MediaType.Carousel, MediaType.Clip, MediaType.IGTV, MediaType.Photo, MediaType.Video]
-        elif media_type is MediaType.Photo:
-            media_type = [MediaType.Carousel, MediaType.Carousel]
-        else:
-            media_type = [media_type]
+        # if media_type is None:
+        #     media_type = [MediaType.Carousel, MediaType.Clip, MediaType.IGTV, MediaType.Photo, MediaType.Video]
+        # elif media_type is MediaType.Photo:
+        #     media_type = [MediaType.Carousel, MediaType.Carousel]
+        # else:
+        #     media_type = [media_type]
 
         tag = tag[1:] if tag[:1] == '#' else tag
         links = []
@@ -33,23 +33,16 @@ class LikeUtil:
 
         return links[:amount]
 
-    def check_media(self, media: Media = None, hashtags_and_phrases_to_skip: list[str] = []) -> bool:
-        media_dict = self.session.media_info(media_pk=media.pk).dict()
-        caption_text = media_dict['caption_text']
+    # check if media is viable for interaction
+    def has_already_liked_media(self, media: Media = None) -> bool:
+        return media.has_liked
 
-        contains_word = False
-        for hapict in hashtags_and_phrases_to_skip:
-            if re.compile(pattern=hapict).search(string=caption_text) is not None:
-                contains_word = True
+    def media_contains_friend(self, media: Media = None, usernames: list[str] = []) -> bool:
+        username = media.user.username
+        return any(friend in username for friend in usernames)
 
-        return not contains_word
+    def media_contains_mandatory_hashtags_or_phrases(self, media: Media = None, hashtags_or_phrases: list[str] = []) -> bool:
+        return True if len(hashtags_or_phrases) == 0 else any(hashtag_or_phrase in media.caption_text for hashtag_or_phrase in hashtags_or_phrases)
 
-
-    def verify_is_likeable(self, media: Media = None, min_likes: int = 0, max_likes: int = 1000) -> bool:
-        if media.has_liked:
-            return False
-
-        return min_likes <= media.like_count <= max_likes
-
-    def like(self, media: Media = None) -> bool:
-        return self.session.media_like(media_id=media.id)
+    def media_contains_hashtags_or_phrases_to_skip(self, media: Media = None, hashtags_or_phrases: list[str] = []) -> bool:
+        return False if len(hashtags_or_phrases) == 0 else any(hashtag_or_phrase in media.caption_text for hashtag_or_phrase in hashtags_or_phrases)