+++ script.py
@@ -6,5 +6,18 @@ import os
 load_dotenv()
 
 session = InstaPy2(username=os.getenv(key='insta_username'), password=os.getenv(key='insta_password'))
-session.set_limit_for_liking(enabled=True, min_likes=0, max_likes=10)
-session.like_by_tags(tags=['python3'], amount=10)
+
+# skips people you add to the list, perfect for skipping friends or less than desirable people (#fuckthem).
+session.set_friends_to_skip(usernames=['...'])
+
+# enable commenting on posts (comment_on_liked_media is unused currently)
+session.set_can_comment(enabled=True, comment_on_liked_media=True, percentage=100)
+
+# sets the comments to use, InstaPy2 will choose a random comment from the list, more comments means more variety.
+session.set_comments(comments=['...']) # use @{} to tag the user of the post
+
+# hashtags or phrases to skip, perfect for nsfw, sex, fuck, etc.
+session.set_hashtags_or_phrases_to_skip(tags=['...'])
+
+# set with tags InstaPy2 will go through with the above configuration.
+session.like_by_tags(tags=['...'], amount=10, skip_top_posts=False)